requirements.txt
anyio==4.0.0
appdirs==1.4.4
astor==0.8.1
asttokens==2.4.1
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
blis==0.7.9
blosc2==2.0.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.4
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Cython==3.0.5
decorator==5.1.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.0
fire==0.4.0
fonttools==4.39.3
formulaic==0.6.6
ftfy==6.1.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
graphlib-backport==1.0.3
h11==0.14.0
h5py==3.6.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.32.0
imbalanced-learn==0.11.0
imblearn==0.0
importlib-resources==5.12.0
interface-meta==1.3.0
ipython==8.12.3
ipywidgets==8.1.1
jdcal==1.4.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.4
langcodes==3.3.0
lifelines==0.27.8
lightning-utilities==0.9.0
lit==16.0.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
murmurhash==1.0.9
networkx==3.1
numexpr==2.8.6
numpy==1.22.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nystrom-attention==0.0.11
opencv-python==4.8.1.78
openpyxl==3.1.2
openslide-python==1.3.1
opt-einsum==3.3.0
packaging==23.1
paddlepaddle==2.5.2
pandas==2.0.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
preshed==3.0.8
progressbar==2.5
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pydantic==1.10.7
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-pretrained-vit==0.0.7
pytz==2023.3
PyYAML==6.0
regex==2023.10.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.33.1
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
slideio==2.2.0
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.3
sympy==1.11.1
tables==3.8.0
termcolor==2.2.0
thinc==8.1.9
threadpoolctl==3.1.0
timm==0.5.4
torch==1.9.1+cu111
torch_geometric==2.4.0
torchaudio==0.9.1
torchmetrics==1.2.0
torchvision==0.10.1+cu111
tqdm==4.65.0
traitlets==5.13.0
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.12
wasabi==1.1.1
wcwidth==0.2.9
widgetsnbextension==4.0.9
wrapt==1.15.0
xlrd==2.0.1
zipp==3.15.0
